+++ CHANGELOG.md
@@ -8,11 +8,25 @@ v3.18.0 - 2018-05-??
 
 + Added a "My Certificates" tab to the Student Dashboard
 + Certificates can be downloaded as HTML files (available when viewing a certificate or from the certificate reporting screen on the admin panel)
++ Admins can now delete certificates from reporting the certificate reporting table on the admin panel
 + Added privacy policy notice on checkout, enrollment, and registration that integrates with the WP Core 4.9.6 Privacy Policy Page setting
 + Added settings to allow customization of the privacy policy and terms & conditions notices during checkout, enrollment, and registration
 + Expanded widths of admin settings page setting names to be a bit wider and more readable
 + Now conditionally hiding some settings when they are no longer relevant
 + Added daily cron automatically remove files from the `LLMS_TMP_DIR` which are more that 24 hours old
+
+##### Privacy Tools
+
++ During a WordPress Personal Data Export request the following LifterLMS information will be added to the export
+  + Student personal information gathered from registration, checkout, and enrollment forms
+  + Student course and membership enrollments, progress, and grades
+  + Student certificates
+  + Student order data
+
++ During a WordPress Personal Data Erasure requet the following LifterLMS information will be erased
+  + Student personal information gathered from registration, checkout, and enrollment forms
+  + Student certificates
+  + If the "Remove Order Data" setting is enabled, the order will be anonymized by removing student personal information from the order and, if the order is a recurring order, it will be cancelled.
 
 ##### Bug Fixes